- We now account for both change messages in the order table making it hopefully correct. We're delayed by a couple minutes on the WebSocket which makes it difficult to tell. Next we should find out if authenticating for the WebSocket gives a better connection - We also updated the navigation to use a BottomNavigationView instead of a drawer that slides in from the left.

+++ app/src/main/java/com/treehouse/gdax/Data/ChangeOrder.kt
@@ -6,7 +6,7 @@ import android.arch.persistence.room.PrimaryKey
 
 @android.arch.persistence.room.Entity(tableName = "CHANGE_ORDERS")
 data class ChangeOrder (
-    @PrimaryKey var sequence: Int = 0,
+    @PrimaryKey var sequence: Long = 0,
     @ColumnInfo var type: String = "",
     @ColumnInfo var time: String = "",
     @ColumnInfo var order_id: String = "",
