+++ app/src/main/java/com/treehouse/gdax/Data/DoneOrder.kt
@@ -7,7 +7,7 @@ import android.arch.persistence.room.PrimaryKey
 
 @android.arch.persistence.room.Entity(tableName = "DONE_ORDERS")
 data class DoneOrder (
-    @PrimaryKey var sequence: Int = 0,
+    @PrimaryKey var sequence: Long = 0,
     @ColumnInfo var type: String = "",
     @ColumnInfo var time: String = "",
     @ColumnInfo var price: String = "",
