+++ app/src/main/java/com/treehouse/gdax/Data/MatchOrder.kt
@@ -6,7 +6,7 @@ import android.arch.persistence.room.PrimaryKey
 
 @Entity(tableName = "MATCH_ORDERS")
 data class MatchOrder (
-    @PrimaryKey var sequence: Int = 0,
+    @PrimaryKey var sequence: Long = 0,
     @ColumnInfo var type: String = "",
     @ColumnInfo var trade_id: Int = 0,
     @ColumnInfo var maker_order_id: String = "",
